11.25

+++ Backend/src/main/java/com/example/Backend/service/ResumeSkillService.java
@@ -22,7 +22,7 @@ public class ResumeSkillService {
         return resumeSkillRepository.save(resumeSkill);
     }
 
-    public void deleteById(Integer resumeId, Integer skillId) {
+    public void deleteById(Integer resumeId, Long  skillId) {
         resumeSkillRepository.deleteById(new ResumeSkillId(resumeId, skillId));
     }
 }
